renamed folder, fixed props
