+++ configurations/nf/synapseConfigs/data.ts
