+++ configurations/nf/synapseConfigs/files.ts
