+++ configurations/nf/synapseConfigs/index.ts
